changed project package structure
