+++ app/src/main/java/com/androidevlinux/percy/UTXO/data/Manager.java
@@ -1,0 +1,4 @@
+package com.androidevlinux.percy.UTXO.data;
+
+public class Manager {
+}
