+++ app/src/main/java/com/androidevlinux/percy/UTXO/data/models/BitfinexPubTickerResponseBean.java
@@ -1,4 +1,4 @@
-package com.androidevlinux.percy.UTXO.Beans;
+package com.androidevlinux.percy.UTXO.data.models;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
