+++ app/src/main/java/com/androidevlinux/percy/UTXO/utils/Constants.java
@@ -1,4 +1,4 @@
-package com.androidevlinux.percy.UTXO.Utils;
+package com.androidevlinux.percy.UTXO.utils;
 
 /**
  * Created by percy on 18/11/17.
