+++ app/src/main/java/com/androidevlinux/percy/UTXO/utils/CustomProgressDialog.java
@@ -1,4 +1,4 @@
-package com.androidevlinux.percy.UTXO.Utils;
+package com.androidevlinux.percy.UTXO.utils;
 
 import android.app.Dialog;
 import android.content.Context;
